A simplified version of minimal toolchains that does the toolchain resolution as the easyconfig deps are being parsed
